extract helper methods used by strategy and auth classes
