+++ packages/shopify-app-remix/src/server/authenticate/admin/helpers/index.ts
@@ -1,7 +1,10 @@
 export * from './begin-auth';
+export * from './create-admin-api-context';
 export * from './handle-client-error';
 export * from './redirect-to-auth-page';
+export * from './redirect-to-shopify-or-app-root';
 export * from './redirect-with-app-bridge-headers';
 export * from './redirect-with-exitiframe';
 export * from './redirect';
 export * from './render-app-bridge';
+export * from './trigger-after-auth-hook';
